refactor(database): split repository.go into service and refresh token files
Separates SqlServiceRepository and SqlRefreshTokenRepository into their own
files within the same package to improve readability and cohesion.
