+++ database/service_repository.go
@@ -2,35 +2,21 @@ package database
 
 import (
 	"context"
-	"crypto/hmac"
-	"crypto/rand"
-	"crypto/sha256"
 	"database/sql"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"strconv"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/fiware/VCVerifier/config"
 	"github.com/fiware/VCVerifier/logging"
 )
 
-// Sentinel errors returned by repository methods.
 var (
 	// ErrServiceNotFound is returned when a service ID does not exist.
 	ErrServiceNotFound = errors.New("service not found")
 	// ErrServiceAlreadyExists is returned on a duplicate service ID insert.
 	ErrServiceAlreadyExists = errors.New("service already exists")
-	// ErrRefreshTokenNotFound is returned when a refresh token does not
-	// exist in the database or has already been consumed.
-	ErrRefreshTokenNotFound = errors.New("refresh token not found")
-	// ErrRefreshTokenInvalidIntegrity is returned when the HMAC of a stored
-	// row does not match, indicating database-level tampering.
-	ErrRefreshTokenInvalidIntegrity = errors.New("refresh token integrity check failed")
 )
 
 // ServiceRepository defines the data-access operations for CCS services
@@ -79,11 +65,6 @@ type SqlServiceRepository struct {
 func NewServiceRepository(db *sql.DB, dbType string) *SqlServiceRepository {
 	return &SqlServiceRepository{db: db, dbType: dbType}
 }
-
-// ---------------------------------------------------------------------------
-// SQL query constants (written with ? placeholders; adapted at runtime for
-// PostgreSQL which requires $N style).
-// ---------------------------------------------------------------------------
 
 const (
 	sqlInsertService = `INSERT INTO service (id, default_oidc_scope, authorization_type) VALUES (?, ?, ?)`
@@ -229,7 +210,6 @@ func (r *SqlServiceRepository) UpdateService(ctx context.Context, id string, ser
 		return config.ConfiguredService{}, fmt.Errorf("update service: %w", err)
 	}
 
-	// Replace all scope entries: delete old, insert new.
 	if _, err := tx.ExecContext(ctx, r.adapt(sqlDeleteScopesByServiceID), id); err != nil {
 		return config.ConfiguredService{}, fmt.Errorf("delete old scopes: %w", err)
 	}
@@ -243,7 +223,6 @@ func (r *SqlServiceRepository) UpdateService(ctx context.Context, id string, ser
 
 	logging.Log().Infof("Updated service %q", id)
 
-	// Re-read from DB to return the persisted state.
 	return r.GetService(ctx, id)
 }
 
@@ -268,7 +247,6 @@ func (r *SqlServiceRepository) DeleteService(ctx context.Context, id string) err
 // GetServiceScopes returns the credential type names required for the given
 // scope. When oidcScope is nil the service's default scope is used.
 func (r *SqlServiceRepository) GetServiceScopes(ctx context.Context, id string, oidcScope *string) ([]string, error) {
-	// Fetch the service to verify existence and resolve default scope.
 	svcRow, err := r.scanServiceRow(ctx, r.db, id)
 	if err != nil {
 		return nil, err
@@ -319,7 +297,7 @@ func (r *SqlServiceRepository) ServiceExists(ctx context.Context, id string) (bo
 }
 
 // ---------------------------------------------------------------------------
-// Internal helpers
+// Internal helpers shared across the database package
 // ---------------------------------------------------------------------------
 
 // queryExecer abstracts *sql.DB and *sql.Tx for shared scan helpers.
@@ -328,8 +306,6 @@ type queryExecer interface {
 	QueryContext(ctx context.Context, query string, args ...interface{}) (*sql.Rows, error)
 }
 
-// scanServiceRow reads a single service row. Returns ErrServiceNotFound when
-// the ID does not exist.
 func (r *SqlServiceRepository) scanServiceRow(ctx context.Context, qe queryExecer, id string) (ServiceRow, error) {
 	var sr ServiceRow
 	err := qe.QueryRowContext(ctx, r.adapt(sqlSelectServiceByID), id).
@@ -343,7 +319,6 @@ func (r *SqlServiceRepository) scanServiceRow(ctx context.Context, qe queryExece
 	return sr, nil
 }
 
-// scanScopeEntryRows reads all scope entries belonging to a service.
 func (r *SqlServiceRepository) scanScopeEntryRows(ctx context.Context, qe queryExecer, serviceID string) ([]ScopeEntryRow, error) {
 	rows, err := qe.QueryContext(ctx, r.adapt(sqlSelectScopesByServiceID), serviceID)
 	if err != nil {
@@ -363,8 +338,6 @@ func (r *SqlServiceRepository) scanScopeEntryRows(ctx context.Context, qe queryE
 	return result, rows.Err()
 }
 
-// insertScopeEntries marshals and inserts scope entries within an existing
-// transaction.
 func (r *SqlServiceRepository) insertScopeEntries(ctx context.Context, tx *sql.Tx, serviceID string, scopes map[string]config.ScopeEntry) error {
 	scopeRows, err := ScopeEntryToRows(serviceID, scopes)
 	if err != nil {
@@ -380,8 +353,6 @@ func (r *SqlServiceRepository) insertScopeEntries(ctx context.Context, tx *sql.T
 	return nil
 }
 
-// batchScopeEntries loads scope entries for multiple service IDs in a single
-// query and groups them by service ID.
 func (r *SqlServiceRepository) batchScopeEntries(ctx context.Context, serviceIDs []string) (map[string][]ScopeEntryRow, error) {
 	if len(serviceIDs) == 0 {
 		return nil, nil
@@ -417,8 +388,6 @@ func (r *SqlServiceRepository) batchScopeEntries(ctx context.Context, serviceIDs
 	return result, rows.Err()
 }
 
-// adapt replaces ? placeholders with $N for PostgreSQL. For MySQL and SQLite
-// the query is returned unchanged.
 func (r *SqlServiceRepository) adapt(query string) string {
 	if r.dbType != DriverTypePostgres {
 		return query
@@ -437,8 +406,6 @@ func (r *SqlServiceRepository) adapt(query string) string {
 	return b.String()
 }
 
-// ph returns the SQL placeholder for the given 1-based parameter position,
-// adapting for the database type ($N for Postgres, ? otherwise).
 func (r *SqlServiceRepository) ph(pos int) string {
 	if r.dbType == DriverTypePostgres {
 		return fmt.Sprintf("$%d", pos)
@@ -453,255 +420,4 @@ func rollbackOnError(tx *sql.Tx) {
 	if err := tx.Rollback(); err != nil && !errors.Is(err, sql.ErrTxDone) {
 		logging.Log().Warnf("rollback failed: %v", err)
 	}
-}
-
-// ---------------------------------------------------------------------------
-// RefreshTokenRepository — database-backed refresh token storage
-// ---------------------------------------------------------------------------
-
-// RefreshTokenRepository defines the data-access operations for OAuth2
-// refresh tokens. Implementations must be safe for concurrent use.
-type RefreshTokenRepository interface {
-	// StoreRefreshToken persists a new refresh token row.
-	StoreRefreshToken(ctx context.Context, row RefreshTokenRow) error
-
-	// GetAndDeleteRefreshToken atomically retrieves and deletes a refresh
-	// token (single-use). Returns ErrRefreshTokenNotFound if the token does
-	// not exist.
-	GetAndDeleteRefreshToken(ctx context.Context, token string) (*RefreshTokenRow, error)
-
-	// DeleteExpiredTokens removes all refresh token rows whose expires_at
-	// is in the past. Returns the number of rows deleted.
-	DeleteExpiredTokens(ctx context.Context) (int64, error)
-
-	// SetCleanupInterval starts a background goroutine that periodically calls
-	// DeleteExpiredTokens at the given interval. If interval is zero or
-	// negative, any running cleanup goroutine is cancelled and no new one is
-	// started. Calling again with a new interval replaces the previous one.
-	// The goroutine stops when ctx is cancelled.
-	SetCleanupInterval(ctx context.Context, interval time.Duration)
-}
-
-// SqlRefreshTokenRepository is a RefreshTokenRepository backed by database/sql.
-type SqlRefreshTokenRepository struct {
-	db            *sql.DB
-	dbType        string
-	mu            sync.Mutex
-	cancelCleanup context.CancelFunc
-	hashEnabled   bool
-	salt          []byte
-}
-
-// NewRefreshTokenRepository creates a new SqlRefreshTokenRepository for the
-// provided database connection and driver type.
-func NewRefreshTokenRepository(db *sql.DB, dbType string) *SqlRefreshTokenRepository {
-	return &SqlRefreshTokenRepository{db: db, dbType: dbType}
-}
-
-// GenerateSalt returns 32 cryptographically random bytes suitable for use
-// with ConfigureHashing.
-func GenerateSalt() ([]byte, error) {
-	salt := make([]byte, 32)
-	if _, err := rand.Read(salt); err != nil {
-		return nil, fmt.Errorf("generate refresh token salt: %w", err)
-	}
-	return salt, nil
-}
-
-// ConfigureHashing enables HMAC-SHA256 hashing of tokens before storage.
-// Must be called before any tokens are stored or retrieved. The salt must
-// not be empty.
-func (r *SqlRefreshTokenRepository) ConfigureHashing(salt []byte) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.salt = salt
-	r.hashEnabled = true
-}
-
-// tokenKey returns the value used as the DB primary key for rawToken:
-// the HMAC-SHA256 hex digest when hashing is enabled, or rawToken as-is.
-func (r *SqlRefreshTokenRepository) tokenKey(rawToken string) string {
-	if !r.hashEnabled {
-		return rawToken
-	}
-	mac := hmac.New(sha256.New, r.salt)
-	mac.Write([]byte(rawToken))
-	return hex.EncodeToString(mac.Sum(nil))
-}
-
-// rawTokenSuffix returns the last 5 characters of token for storage.
-func rawTokenSuffix(token string) string {
-	if len(token) <= 5 {
-		return token
-	}
-	return token[len(token)-5:]
-}
-
-// SQL query constants for refresh token operations (written with ?
-// placeholders; adapted at runtime for PostgreSQL).
-const (
-	sqlInsertRefreshToken = `INSERT INTO refresh_token (token, token_suffix, client_id, claims, integrity, expires_at) VALUES (?, ?, ?, ?, ?, ?)`
-
-	sqlSelectRefreshToken = `SELECT token, token_suffix, client_id, claims, integrity, expires_at FROM refresh_token WHERE token = ?`
-
-	sqlDeleteRefreshToken = `DELETE FROM refresh_token WHERE token = ?`
-
-	sqlDeleteExpiredRefreshTokens = `DELETE FROM refresh_token WHERE expires_at < ?`
-)
-
-// computeIntegrity returns HMAC-SHA256(salt, rawToken|"|"|clientId|"|"|claims)
-// as a lowercase hex string. Returns empty string when salt is not configured,
-// in which case integrity verification is skipped on retrieval.
-func computeIntegrity(salt []byte, rawToken, clientId, claims string, expiresAt int64) string {
-	if len(salt) == 0 {
-		return ""
-	}
-	mac := hmac.New(sha256.New, salt)
-	mac.Write([]byte(rawToken))
-	mac.Write([]byte("|"))
-	mac.Write([]byte(clientId))
-	mac.Write([]byte("|"))
-	mac.Write([]byte(claims))
-	mac.Write([]byte("|"))
-	mac.Write([]byte(strconv.FormatInt(expiresAt, 10)))
-	return hex.EncodeToString(mac.Sum(nil))
-}
-
-// StoreRefreshToken persists a new refresh token row in the database.
-// The token primary key is hashed when hashing is configured; the last 5
-// characters of the raw token are always stored in token_suffix. The integrity
-// HMAC is computed from the configured salt and stored alongside the claims.
-func (r *SqlRefreshTokenRepository) StoreRefreshToken(ctx context.Context, row RefreshTokenRow) error {
-	integrity := computeIntegrity(r.salt, row.Token, row.ClientID, row.Claims, row.ExpiresAt)
-	_, err := r.db.ExecContext(ctx, r.adapt(sqlInsertRefreshToken),
-		r.tokenKey(row.Token), rawTokenSuffix(row.Token), row.ClientID, row.Claims, integrity, row.ExpiresAt)
-	if err != nil {
-		return fmt.Errorf("insert refresh token: %w", err)
-	}
-	logging.Log().Debugf("Stored refresh token (expires_at=%d)", row.ExpiresAt)
-	return nil
-}
-
-// GetAndDeleteRefreshToken atomically retrieves and deletes a refresh token
-// within a transaction, ensuring single-use semantics. Returns
-// ErrRefreshTokenNotFound if the token does not exist, or
-// ErrRefreshTokenInvalidIntegrity if the stored HMAC does not match (possible
-// database-level tampering). The token is consumed regardless of the integrity
-// outcome to prevent repeated use of a tampered row.
-func (r *SqlRefreshTokenRepository) GetAndDeleteRefreshToken(ctx context.Context, token string) (*RefreshTokenRow, error) {
-	tx, err := r.db.BeginTx(ctx, nil)
-	if err != nil {
-		return nil, fmt.Errorf("begin tx: %w", err)
-	}
-	defer rollbackOnError(tx)
-
-	var row RefreshTokenRow
-	err = tx.QueryRowContext(ctx, r.adapt(sqlSelectRefreshToken), r.tokenKey(token)).Scan(
-		&row.Token, &row.TokenSuffix, &row.ClientID, &row.Claims, &row.Integrity, &row.ExpiresAt)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			return nil, ErrRefreshTokenNotFound
-		}
-		return nil, fmt.Errorf("select refresh token: %w", err)
-	}
-
-	if _, err := tx.ExecContext(ctx, r.adapt(sqlDeleteRefreshToken), r.tokenKey(token)); err != nil {
-		return nil, fmt.Errorf("delete refresh token: %w", err)
-	}
-
-	if err := tx.Commit(); err != nil {
-		return nil, fmt.Errorf("commit: %w", err)
-	}
-
-	// Verify integrity after committing so the token is consumed even on failure,
-	// preventing repeated use of a tampered row.
-	if len(r.salt) > 0 {
-		expected := computeIntegrity(r.salt, token, row.ClientID, row.Claims, row.ExpiresAt)
-		if !hmac.Equal([]byte(expected), []byte(row.Integrity)) {
-			logging.Log().Warnf("Refresh token integrity check failed (suffix=%s): possible database-level tampering", row.TokenSuffix)
-			return nil, ErrRefreshTokenInvalidIntegrity
-		}
-	}
-
-	logging.Log().Debugf("Retrieved and deleted refresh token")
-	return &row, nil
-}
-
-// DeleteExpiredTokens removes all refresh token rows whose expiration time
-// has passed. Returns the number of rows deleted.
-func (r *SqlRefreshTokenRepository) DeleteExpiredTokens(ctx context.Context) (int64, error) {
-	now := time.Now().Unix()
-	result, err := r.db.ExecContext(ctx, r.adapt(sqlDeleteExpiredRefreshTokens), now)
-	if err != nil {
-		return 0, fmt.Errorf("delete expired refresh tokens: %w", err)
-	}
-	n, err := result.RowsAffected()
-	if err != nil {
-		return 0, fmt.Errorf("rows affected: %w", err)
-	}
-	if n > 0 {
-		logging.Log().Infof("Cleaned up %d expired refresh token(s)", n)
-	}
-	return n, nil
-}
-
-// SetCleanupInterval starts a background goroutine that periodically deletes
-// expired refresh token rows. If interval is zero or negative, any running
-// cleanup goroutine is cancelled and no new one is started. Calling again
-// replaces the previous interval. The goroutine stops when ctx is cancelled.
-func (r *SqlRefreshTokenRepository) SetCleanupInterval(ctx context.Context, interval time.Duration) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	if r.cancelCleanup != nil {
-		r.cancelCleanup()
-		r.cancelCleanup = nil
-	}
-
-	if interval <= 0 {
-		return
-	}
-
-	cleanupCtx, cancel := context.WithCancel(ctx)
-	r.cancelCleanup = cancel
-
-	go func() {
-		runCleanup := func() {
-			logging.Log().Debug("Running refresh token cleanup")
-			if _, err := r.DeleteExpiredTokens(cleanupCtx); err != nil && cleanupCtx.Err() == nil {
-				logging.Log().Warnf("Refresh token cleanup failed: %v", err)
-			}
-		}
-		runCleanup()
-		ticker := time.NewTicker(interval)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ticker.C:
-				runCleanup()
-			case <-cleanupCtx.Done():
-				return
-			}
-		}
-	}()
-}
-
-// adapt replaces ? placeholders with $N for PostgreSQL. For MySQL and SQLite
-// the query is returned unchanged.
-func (r *SqlRefreshTokenRepository) adapt(query string) string {
-	if r.dbType != DriverTypePostgres {
-		return query
-	}
-	var b strings.Builder
-	b.Grow(len(query))
-	n := 1
-	for i := 0; i < len(query); i++ {
-		if query[i] == '?' {
-			fmt.Fprintf(&b, "$%d", n)
-			n++
-		} else {
-			b.WriteByte(query[i])
-		}
-	}
-	return b.String()
 }
